Revert "chore: cleanup stale resources in CodeBuild (#12779)"
This reverts commit de9865416e5de19e3dfd807e70f1bc34683058aa.
