+++ codebuild_specs/cleanup_resources.yml
@@ -4,9 +4,4 @@ env:
 phases:
   build:
     commands:
-      - cd ./scripts
-      - npm install -g ts-node
-      - npm install aws-sdk
-      - ts-node ./wait-for-all-codebuild.ts $CODEBUILD_RESOLVED_SOURCE_VERSION ../$WAIT_FOR_IDS_FILE_PATH AmplifyCLI-E2E-Testing
-      - cd ..
-      - source ./shared-scripts.sh && _cleanUpResources
+      - echo cleanup running
