+++ codebuild_specs/e2e_workflow.yml
@@ -149,11 +149,3 @@ batch:
       debug-session: true
       depend-on:
         - upload_pkg_binaries
-    - identifier: cleanup_resources
-      buildspec: codebuild_specs/cleanup_resources.yml
-      env:
-        compute-type: BUILD_GENERAL1_LARGE
-        variables:
-          WAIT_FOR_IDS_FILE_PATH: ./codebuild_specs/wait_for_ids_local.json
-      depend-on:
-        - upload_pkg_binaries
