+++ packages/amplify-e2e-core/src/init/initProjectHelper.ts
@@ -3,7 +3,7 @@
 /* eslint-disable func-style */
 import { EOL } from 'os';
 import { v4 as uuid } from 'uuid';
-import { nspawn as spawn, getCLIPath, singleSelect, addCICleanupTags } from '..';
+import { nspawn as spawn, getCLIPath, singleSelect, addCircleCITags } from '..';
 import { KEY_DOWN_ARROW } from '../utils';
 import { amplifyRegions } from '../configure';
 
@@ -39,7 +39,7 @@ export function initJSProjectWithProfile(cwd: string, settings?: Partial<typeof 
     };
   }
 
-  addCICleanupTags(cwd);
+  addCircleCITags(cwd);
 
   const cliArgs = ['init'];
   const providerConfigSpecified = !!s.providerConfig && typeof s.providerConfig === 'object';
@@ -98,7 +98,7 @@ export function initJSProjectWithProfile(cwd: string, settings?: Partial<typeof 
 export function initAndroidProjectWithProfile(cwd: string, settings: Partial<typeof defaultSettings>): Promise<void> {
   const s = { ...defaultSettings, ...settings };
 
-  addCICleanupTags(cwd);
+  addCircleCITags(cwd);
 
   let env;
 
@@ -135,7 +135,7 @@ export function initAndroidProjectWithProfile(cwd: string, settings: Partial<typ
       .wait(/Try "amplify add api" to create a backend API and then "amplify (push|publish)" to deploy everything/)
       .run((err: Error) => {
         if (!err) {
-          addCICleanupTags(cwd);
+          addCircleCITags(cwd);
 
           resolve();
         } else {
@@ -154,7 +154,7 @@ export function createRandomName(): string {
 export function initIosProjectWithProfile(cwd: string, settings: Record<string, unknown>): Promise<void> {
   const s = { ...defaultSettings, ...settings };
 
-  addCICleanupTags(cwd);
+  addCircleCITags(cwd);
 
   let env;
 
@@ -189,7 +189,7 @@ export function initIosProjectWithProfile(cwd: string, settings: Record<string, 
       .wait(/Try "amplify add api" to create a backend API and then "amplify (push|publish)" to deploy everything/)
       .run((err: Error) => {
         if (!err) {
-          addCICleanupTags(cwd);
+          addCircleCITags(cwd);
 
           resolve();
         } else {
@@ -202,7 +202,7 @@ export function initIosProjectWithProfile(cwd: string, settings: Record<string, 
 export function initFlutterProjectWithProfile(cwd: string, settings: Record<string, unknown>): Promise<void> {
   const s = { ...defaultSettings, ...settings };
 
-  addCICleanupTags(cwd);
+  addCircleCITags(cwd);
 
   return new Promise((resolve, reject) => {
     const chain = spawn(getCLIPath(), ['init'], { cwd, stripColors: true })
@@ -245,7 +245,7 @@ export function initProjectWithAccessKey(
 ): Promise<void> {
   const s = { ...defaultSettings, ...settings };
 
-  addCICleanupTags(cwd);
+  addCircleCITags(cwd);
 
   return new Promise((resolve, reject) => {
     const chain = spawn(getCLIPath(), ['init'], {
@@ -303,7 +303,7 @@ export function initProjectWithAccessKey(
 }
 
 export function initNewEnvWithAccessKey(cwd: string, s: { envName: string; accessKeyId: string; secretAccessKey: string }): Promise<void> {
-  addCICleanupTags(cwd);
+  addCircleCITags(cwd);
 
   return new Promise((resolve, reject) => {
     const chain = spawn(getCLIPath(), ['init'], {
@@ -341,7 +341,7 @@ export function initNewEnvWithAccessKey(cwd: string, s: { envName: string; acces
 }
 
 export function initNewEnvWithProfile(cwd: string, s: { envName: string }): Promise<void> {
-  addCICleanupTags(cwd);
+  addCircleCITags(cwd);
 
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['init'], {
@@ -372,7 +372,7 @@ export function initNewEnvWithProfile(cwd: string, s: { envName: string }): Prom
 }
 
 export function updatedInitNewEnvWithProfile(cwd: string, s: { envName: string }): Promise<void> {
-  addCICleanupTags(cwd);
+  addCircleCITags(cwd);
 
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['init'], {
@@ -409,7 +409,7 @@ export function amplifyInitSandbox(cwd: string, settings: Record<string, unknown
     };
   }
 
-  addCICleanupTags(cwd);
+  addCircleCITags(cwd);
 
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['init'], { cwd, stripColors: true, env })
