+++ packages/amplify-e2e-core/src/utils/index.ts
@@ -7,7 +7,7 @@ import execa from 'execa';
 import { v4 as uuid } from 'uuid';
 import { getLayerDirectoryName, LayerDirectoryType } from '..';
 
-export * from './add-ci-cleanup-tags';
+export * from './add-circleci-tags';
 export * from './api';
 export * from './appsync';
 export * from './envVars';
