+++ packages/amplify-e2e-core/src/utils/pinpoint.ts
@@ -1,7 +1,7 @@
 import { Pinpoint } from 'aws-sdk';
 import _ from 'lodash';
 import { EOL } from 'os';
-import { getCLIPath, nspawn as spawn, singleSelect, amplifyRegions, addCICleanupTags, KEY_DOWN_ARROW } from '..';
+import { getCLIPath, nspawn as spawn, singleSelect, amplifyRegions, addCircleCITags, KEY_DOWN_ARROW } from '..';
 
 const settings = {
   name: EOL,
@@ -57,7 +57,7 @@ export async function pinpointAppExist(pinpointProjectId: string, region: string
  * initializes a project to test pinpoint
  */
 export function initProjectForPinpoint(cwd: string): Promise<void> {
-  addCICleanupTags(cwd);
+  addCircleCITags(cwd);
 
   return new Promise((resolve, reject) => {
     const chain = spawn(getCLIPath(), ['init'], {
