+++ packages/amplify-e2e-tests/src/cleanup-e2e-resources.ts
@@ -149,7 +149,7 @@ const STALE_DURATION_MS = 2 * 60 * 60 * 1000; // 2 hours in milliseconds
  */
 const handleExpiredTokenException = (): void => {
   console.log('Token expired. Exiting...');
-  process.exit(1);
+  process.exit();
 };
 
 /**
