+++ packages/amplify-migration-tests/src/migration-helpers-v10/init.ts
@@ -1,4 +1,4 @@
-import { addCICleanupTags, getCLIPath, nspawn as spawn } from '@aws-amplify/amplify-e2e-core';
+import { addCircleCITags, getCLIPath, nspawn as spawn } from '@aws-amplify/amplify-e2e-core';
 import { EOL } from 'os';
 
 const defaultSettings = {
@@ -32,7 +32,7 @@ export function initJSProjectWithProfileV10(cwd: string, settings?: Partial<type
     };
   }
 
-  addCICleanupTags(cwd);
+  addCircleCITags(cwd);
 
   const cliArgs = ['init'];
   const providerConfigSpecified = !!s.providerConfig && typeof s.providerConfig === 'object';
