+++ scripts/package.json
@@ -19,25 +19,18 @@
     "@types/js-yaml": "^4.0.4",
     "@types/node": "^16.11.5",
     "@yarnpkg/parsers": "^3.0.0-rc.44",
-    "aws-sdk": "^2.1354.0",
     "axios": "^0.26.0",
     "execa": "^5.1.1",
     "fs-extra": "^10.0.0",
     "glob": "^7.2.0",
     "js-yaml": "^4.1.0",
-    "lodash": "^4.17.21",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.5",
-    "yargs": "^17.7.2"
+    "typescript": "^4.9.5"
   },
   "berry": {
     "plugins": [
       "@yarn/plugin-typescript"
     ]
   },
-  "packageManager": "yarn@3.5.0",
-  "devDependencies": {
-    "@types/lodash": "^4",
-    "@types/yargs": "^17"
-  }
+  "packageManager": "yarn@3.5.0"
 }
